# Speedup tests