Benchmarked with `hyperfine`. Not patched:

```
⇒  hyperfine pytest
Benchmark #1: pytest
  Time (mean ± σ):     26.637 s ±  0.321 s    [User: 22.430 s, System: 0.938 s]
  Range (min … max):   26.213 s … 27.140 s    10 runs
```

vs patched:

```
⇒  hyperfine pytest   
Benchmark #1: pytest
  Time (mean ± σ):     24.258 s ±  0.273 s    [User: 20.051 s, System: 0.958 s]
  Range (min … max):   23.969 s … 24.913 s    10 run
```

which is ~9% improvement